backend/pyproject.toml
[tool.poetry]
name = "capstone"
version = "0.1.0"
description = "Final project for CS50W"
authors = ["rodvicj <rodvicj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.4"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
django-markdownx = "^4.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
